fix growthAffectingPolymerases; newer matplotlib
* growthAffectingPolymerases: Adapt "TypeError: `bins` must be an integer, a string, or an array" for recent pips.

* Update to matplotlib==1.5.3. This fixes "FutureWarning: elementwise comparison failed; returning scalar instead, but in the future will perform elementwise comparison".

+++ models/ecoli/analysis/multigen/growthAffectingPolymerases.py
@@ -14,6 +14,7 @@ import matplotlib.gridspec as gridspec
 from models.ecoli.analysis.AnalysisPaths import AnalysisPaths
 from wholecell.io.tablereader import TableReader
 import wholecell.utils.constants
+from wholecell.analysis.analysis_tools import exportFigure
 from wholecell.utils import units
 import cPickle
 
@@ -219,7 +220,7 @@ def main(seedOutDir, plotOutDir, plotOutFileName, simDataFile, validationDataFil
 		ax1.axvline(x = time.asNumber(units.min).max(), linewidth=2, color='k', linestyle='--')
 
 		hist_doublingTime = removeNanReshape(doublingTime.asNumber(units.min))
-		nbins = np.ceil(np.sqrt(hist_doublingTime.size))
+		nbins = int(np.ceil(np.sqrt(hist_doublingTime.size)))
 		ax1_1.hist(hist_doublingTime, nbins, (hist_doublingTime.mean() - hist_doublingTime.std() / 2, hist_doublingTime.mean() + hist_doublingTime.std() / 2))
 
 		# Plot RNAP active fraction
@@ -229,7 +230,7 @@ def main(seedOutDir, plotOutDir, plotOutFileName, simDataFile, validationDataFil
 		ax2.set_ylabel("Fraction active\nRNAP")
 
 		hist_fractionRnapActive = removeNanReshape(fractionRnapActive)
-		nbins = np.ceil(np.sqrt(hist_fractionRnapActive.size))
+		nbins = int(np.ceil(np.sqrt(hist_fractionRnapActive.size)))
 		ax2_1.hist(hist_fractionRnapActive, nbins, (hist_fractionRnapActive.mean() - hist_fractionRnapActive.std(), hist_fractionRnapActive.mean() + hist_fractionRnapActive.std()))
 
 		# Plot RNAP active and total counts
@@ -359,13 +360,12 @@ def main(seedOutDir, plotOutDir, plotOutFileName, simDataFile, validationDataFil
 
 	fig.subplots_adjust(hspace=.5, wspace = 0.3)
 
-	from wholecell.analysis.analysis_tools import exportFigure
 	exportFigure(plt, plotOutDir, plotOutFileName,metadata)
 	plt.close("all")
 
 def getMassData(simDir, massNames):
 	simOutDir = os.path.join(simDir, "simOut")
-	
+
 	time = TableReader(os.path.join(simOutDir, "Main")).readColumn("time")
 
 	mass = TableReader(os.path.join(simOutDir, "Mass"))
